backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
django-filter==25.1
djangorestframework-simplejwt==5.5.1
django-guardian==3.1.0
Pillow==11.3.0
